https://www.youtube.com/watch?v=ldJ8WGZVXZk&pp=ygUMdGhyZWFkcyBpbiBj
The pthread_t data type is used to uniquely identify and manage threads.

Two threads (thread1 and thread2) are created using pthread_create. Each thread runs the computation function, receiving a different value as an argument.

The computation function performs a loop to simulate computational work. The value passed to the function is added to a sum variable in each iteration of the loop.

The program uses pthread_join to wait for both threads to finish their execution before the program exits.

The computation performed by each thread is independent, allowing for potential parallel execution.